small updates.
sum up what's left to do:

Difference - fill in and javadoc
product - fill in and Javadoc
Quotient - fill in and Javadoc
Constant - Javadoc
MultiSum - Javadoc

test code

+++ src/PartB/Function.java
@@ -10,7 +10,7 @@ public abstract class Function {
     public abstract Function derivative();
     public double bisectionMethod(double a, double b, double epsilon){
         double left=a, right = b;
-        while((double)(right - left) > epsilon){
+        while((right - left) > epsilon){
             double mid = (left+right)/2;
             if(this.valueAt(left) * this.valueAt(mid) > 0)
                 left = mid;
